Added search for profile
Just realized that the filter for loading more content (if there is such) relies on the front end's search box which means that the consistency of the search will fail if the content of the search box for the specific page that the button is attached to is empty or changed in value.

Still it just a UX issue than one that affects the functionality. Though fixing said issue requires rework on the backend side as well to compensate for the handling of the search filters.